Track command names in log and "bot stats"
