+++ bot/bot_handle_test.go
@@ -3,73 +3,9 @@ package bot
 import (
 	"testing"
 
-	"github.com/innogames/slack-bot/v2/bot/config"
-	"github.com/innogames/slack-bot/v2/bot/matcher"
-	"github.com/innogames/slack-bot/v2/bot/msg"
-	"github.com/innogames/slack-bot/v2/client"
 	"github.com/slack-go/slack"
 	"github.com/stretchr/testify/assert"
 )
-
-func TestBot(t *testing.T) {
-	cfg := config.Config{}
-	cfg.Slack.Token = "xoxb-12345"
-	cfg.AdminUsers = config.UserList{
-		"admin123",
-	}
-
-	slackClient, err := client.GetSlackClient(cfg.Slack)
-	assert.Nil(t, err)
-
-	commands := &Commands{}
-	commands.AddCommand(testCommand2{})
-
-	bot := NewBot(cfg, slackClient, commands)
-	bot.auth = &slack.AuthTestResponse{
-		UserID: "BOT",
-	}
-	bot.allowedUsers = map[string]string{
-		"U123": "User123",
-	}
-
-	t.Run("handle empty message", func(t *testing.T) {
-		message := msg.Message{}
-		message.Text = ""
-		message.Channel = "C123"
-		bot.ProcessMessage(message, true)
-	})
-
-	t.Run("handle unauthenticated message", func(t *testing.T) {
-		message := msg.Message{}
-
-		message.Text = "test"
-		message.User = "U888"
-		message.Channel = "C123"
-		bot.ProcessMessage(message, true)
-	})
-
-	t.Run("handle valid message", func(t *testing.T) {
-		message := msg.Message{}
-
-		message.Text = "test"
-		message.User = "U123"
-		message.Channel = "C123"
-		bot.ProcessMessage(message, true)
-	})
-
-	t.Run("Init with invalid token", func(t *testing.T) {
-		bot.config.Slack.Token = "invalid"
-
-		err := bot.Init()
-		assert.EqualError(t, err, "auth error: invalid_auth")
-	})
-
-	t.Run("Load channels without token", func(t *testing.T) {
-		channels, err := bot.loadChannels()
-		assert.Len(t, channels, 0)
-		assert.Contains(t, err.Error(), "invalid_auth")
-	})
-}
 
 func TestIsBotMessage(t *testing.T) {
 	bot := Bot{}
@@ -159,12 +95,6 @@ func BenchmarkTrimMessage(b *testing.B) {
 	})
 }
 
-func TestInitLogger(t *testing.T) {
-	cfg := config.Config{}
-	cfg.Logger.Level = "debug"
-	InitLogger(cfg.Logger)
-}
-
 func BenchmarkShouldHandle(b *testing.B) {
 	bot := Bot{}
 	bot.auth = &slack.AuthTestResponse{}
@@ -181,12 +111,5 @@ func BenchmarkShouldHandle(b *testing.B) {
 			result = bot.canHandleMessage(event)
 		}
 		assert.True(b, result)
-	})
-}
-
-type testCommand2 struct{}
-
-func (c testCommand2) GetMatcher() matcher.Matcher {
-	return matcher.NewTextMatcher("test", func(match matcher.Result, message msg.Message) {
 	})
 }
