+++ bot/command.go
@@ -1,6 +1,9 @@
 package bot
 
 import (
+	"reflect"
+	"sort"
+	"strings"
 	"sync"
 
 	"github.com/innogames/slack-bot/v2/bot/matcher"
@@ -39,9 +42,10 @@ type HelpProvider interface {
 
 // Commands is a wrapper of a list of commands. Only the first matched command will be executed
 type Commands struct {
-	commands []Command
-	matcher  []matcher.Matcher // precompiled matcher objects
-	compiled bool
+	commands     []Command
+	matcher      []matcher.Matcher // precompiled matcher objects
+	matcherNames map[int]string    // precompiled mapping from matcher -> command name
+	compiled     bool
 }
 
 // GetHelp returns the help for ALL included commands
@@ -59,9 +63,16 @@ func (c *Commands) GetHelp() []Help {
 
 // Run executes the first matched command and return true in case one command matched
 func (c *Commands) Run(message msg.Message) bool {
+	matched, _ := c.RunWithName(message)
+
+	return matched
+}
+
+// RunWithName executes the first matched command and return the command name if there was a match
+func (c *Commands) RunWithName(message msg.Message) (bool, string) {
 	c.compile()
 
-	for _, command := range c.matcher {
+	for i, command := range c.matcher {
 		run, match := command.Match(message)
 		if match != nil {
 			// this is is needed for ConditionMatcher: runner gets already executed in the matcher itself!
@@ -70,11 +81,11 @@ func (c *Commands) Run(message msg.Message) bool {
 			}
 
 			// only the first command is executed -> abort here
-			return true
+			return true, c.matcherNames[i]
 		}
 	}
 
-	return false
+	return false, ""
 }
 
 // AddCommand registers a command to the command list
@@ -109,7 +120,22 @@ func (c *Commands) Merge(commands Commands) {
 
 // Count the registered/valid commands
 func (c *Commands) Count() int {
+	c.compile()
+
 	return len(c.commands)
+}
+
+func (c *Commands) GetCommandNames() []string {
+	c.compile()
+
+	names := make([]string, 0, len(c.matcherNames))
+	for _, name := range c.matcherNames {
+		names = append(names, name)
+	}
+
+	sort.Strings(names)
+
+	return names
 }
 
 func (c *Commands) compile() {
@@ -123,9 +149,24 @@ func (c *Commands) compile() {
 
 	if !c.compiled {
 		c.matcher = make([]matcher.Matcher, len(c.commands))
+		c.matcherNames = make(map[int]string, len(c.commands))
+
 		for i, command := range c.commands {
-			c.matcher[i] = command.GetMatcher()
+			commandName := getCommandName(command)
+			commandMatcher := command.GetMatcher()
+
+			c.matcher[i] = commandMatcher
+			c.matcherNames[i] = commandName
 		}
 		c.compiled = true
 	}
+}
+
+func getCommandName(command Command) string {
+	t := reflect.TypeOf(command)
+	name := t.String()
+	name = strings.ReplaceAll(name, "*", "")
+	name = strings.TrimSuffix(name, "Command")
+
+	return name
 }
