+++ bot/config/dump_test.go
@@ -10,6 +10,6 @@ func TestDump(t *testing.T) {
 	cfg := DefaultConfig
 	actual := Dump(cfg)
 
-	assert.Contains(t, actual, `"Timezone": ""`)
-	assert.Contains(t, actual, `"File": "./bot.log"`)
+	assert.Contains(t, actual, `timezone: ""`)
+	assert.Contains(t, actual, `debug: false`)
 }
